fix: 명확한 도메인 이름 사용
